refactored get vehicle make model to tech record repository
