+++ tests/unit/CertificateGenerationService.unitTest.ts
@@ -9,19 +9,12 @@ import { cloneDeep } from 'lodash';
 import Axios from 'axios';
 import { toUint8Array } from '@smithy/util-utf8';
 import { CertificateGenerationService } from '../../src/services/CertificateGenerationService';
-import techRecordResp from '../resources/tech-records-response.json';
 import testResultsResp from '../resources/test-results-response.json';
 import testResultsRespFail from '../resources/test-results-fail-response.json';
 import testResultsRespPrs from '../resources/test-results-prs-response.json';
 import testResultsRespEmpty from '../resources/test-results-empty-response.json';
 import testResultsRespNoCert from '../resources/test-results-nocert-response.json';
 import { LambdaService } from '../../src/services/LambdaService';
-import techRecordsRwtSearch from '../resources/tech-records-response-rwt-search.json';
-import techRecordsRwt from '../resources/tech-records-response-rwt.json';
-import techRecordsRwtHgv from '../resources/tech-records-response-rwt-hgv.json';
-import techRecordsRwtHgvSearch from '../resources/tech-records-response-rwt-hgv-search.json';
-import techRecordsPsv from '../resources/tech-records-response-PSV.json';
-import techRecordsSearchPsv from '../resources/tech-records-response-search-PSV.json';
 import mockTestResult from '../resources/test-result-with-defect.json';
 import defectsMock from '../resources/defects_mock.json';
 import flatDefectsMock from '../resources/flattened-defects.json';
@@ -62,220 +55,6 @@ describe('Certificate Generation Service', () => {
     searchTechRecordsSpy.mockReset();
     callGetTechRecordSpy.mockReset();
     invokeSpy.mockReset();
-  });
-
-  describe('getVehicleMakeAndModel function', () => {
-    context('when given a systemNumber with matching record', () => {
-      it('should return the record & only invoke the LambdaService once', async () => {
-        const certGenSvc = Container.get(CertificateGenerationService);
-        searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-
-        const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-        callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-        const testResultMock = {
-          systemNumber: '12345678',
-        };
-        const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-          testResultMock,
-        );
-        expect(makeAndModel).toEqual({ Make: 'STANLEY', Model: 'AUTOTRL' });
-      });
-    });
-
-    context(
-      'when given a systemNumber  with no matching record and a vin with matching record',
-      () => {
-        it('should return the record & invoke the LambdaService twice', async () => {
-          const certGenSvc = Container.get(CertificateGenerationService);
-          searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-
-          const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-          callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-          const testResultMock = {
-            systemNumber: '134567889',
-            vin: 'abc123',
-          };
-          const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-            testResultMock,
-          );
-          expect(makeAndModel).toEqual({ Make: 'STANLEY', Model: 'AUTOTRL' });
-        });
-      },
-    );
-
-    context(
-      'when given a vin with no matching record but a matching partialVin',
-      () => {
-        it('should return the record & invoke the LambdaService twice', async () => {
-          invokeSpy
-            .mockResolvedValueOnce(AWSReject('no'))
-            .mockResolvedValueOnce(AWSResolve(JSON.stringify(techRecordResp)));
-
-          const certGenSvc = Container.get(CertificateGenerationService);
-          searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-
-          const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-          callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-          const testResultMock = {
-            vin: 'abc123',
-            partialVin: 'abc123',
-          };
-          const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-            testResultMock,
-          );
-          expect(makeAndModel).toEqual({ Make: 'STANLEY', Model: 'AUTOTRL' });
-        });
-      },
-    );
-
-    context(
-      'when given a vin and partialVin with no matching record but a matching VRM',
-      () => {
-        it('should return the record & invoke the LambdaService three times', async () => {
-          invokeSpy
-            .mockResolvedValueOnce(AWSReject('no'))
-            .mockResolvedValueOnce(AWSReject('no'))
-            .mockResolvedValueOnce(AWSResolve(JSON.stringify(techRecordResp)));
-
-          const certGenSvc = Container.get(CertificateGenerationService);
-          searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-          const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-          callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-          const testResultMock = {
-            vin: 'abc123',
-            partialVin: 'abc123',
-            vrm: 'testvrm',
-          };
-          const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-            testResultMock,
-          );
-          expect(invokeSpy).not.toHaveBeenCalledTimes(1);
-          expect(invokeSpy).not.toHaveBeenCalledTimes(2);
-          expect(makeAndModel).toEqual({ Make: 'STANLEY', Model: 'AUTOTRL' });
-        });
-      },
-    );
-
-    context(
-      'when given a vin, partialVin and VRM with no matching record but a matching TrailerID',
-      () => {
-        it('should return the record & invoke the LambdaService four times', async () => {
-          const certGenSvc = Container.get(CertificateGenerationService);
-          searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-
-          const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-          callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-          const testResultMock = {
-            vin: 'abc123',
-            partialVin: 'abc123',
-            vrm: 'testvrm',
-            trailerId: 'testTrailerId',
-          };
-          const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-            testResultMock,
-          );
-          expect(makeAndModel).toEqual({ Make: 'STANLEY', Model: 'AUTOTRL' });
-        });
-      },
-    );
-
-    context('when given a vehicle details with no matching records', () => {
-      it('should call Tech Records Lambda 4 times and then throw an error', async () => {
-        invokeSpy
-          .mockResolvedValue(AWSReject('no'));
-        const certGenSvc = Container.get(CertificateGenerationService);
-        searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-
-        const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-        callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-        const testResultMock = {
-          vin: 'abc123',
-          partialVin: 'abc123',
-          vrm: 'testvrm',
-          trailerId: 'testTrailerId',
-        };
-        try {
-          await certGenSvc.getVehicleMakeAndModel(testResultMock);
-        } catch (e) {
-          expect(invokeSpy).toBeCalledTimes(4);
-          expect(e).toBeInstanceOf(Error);
-          expect((e as Error).message).toBe(
-            'Unable to retrieve unique Tech Record for Test Result',
-          );
-        }
-      });
-    });
-
-    context(
-      'when given a vehicle details with missing vehicle detail fields and no match',
-      () => {
-        it('should call Tech Records Lambda matching number (2) times and then throw an error', async () => {
-          invokeSpy
-            .mockResolvedValue(AWSReject('no'));
-          const certGenSvc = Container.get(CertificateGenerationService);
-          searchTechRecordsSpy.mockResolvedValue(techRecordsRwtSearch);
-
-          const techRecordResponseRwtMock = cloneDeep(techRecordsRwt);
-          callGetTechRecordSpy.mockResolvedValue(techRecordResponseRwtMock as any);
-
-          const testResultMock = {
-            vin: 'abc123',
-            trailerId: 'testTrailerId',
-          };
-          try {
-            await certGenSvc.getVehicleMakeAndModel(testResultMock);
-          } catch (e) {
-            expect(invokeSpy).toBeCalledTimes(2);
-            expect(e).toBeInstanceOf(Error);
-            expect((e as Error).message).toBe(
-              'Unable to retrieve unique Tech Record for Test Result',
-            );
-          }
-        });
-      },
-    );
-
-    context('when lookup returns a PSV tech record', () => {
-      it('should return make and model from chassis details', async () => {
-        const certGenSvc = Container.get(CertificateGenerationService);
-        searchTechRecordsSpy.mockResolvedValue(techRecordsSearchPsv);
-
-        callGetTechRecordSpy.mockResolvedValue(techRecordsPsv as any);
-
-        const testResultMock = {
-          systemNumber: '12345678',
-        };
-        const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-          testResultMock,
-        );
-        expect(makeAndModel.Make).toBe('AEC');
-        expect(makeAndModel.Model).toBe('RELIANCE');
-      });
-    });
-
-    context('when lookup returns a non-PSV tech record', () => {
-      it('should return make and model from not-chassis details', async () => {
-        const certGenSvc = Container.get(CertificateGenerationService);
-        searchTechRecordsSpy.mockResolvedValue(techRecordsRwtHgvSearch);
-
-        callGetTechRecordSpy.mockResolvedValue(techRecordsRwtHgv as any);
-
-        const testResultMock = {
-          systemNumber: '12345678',
-        };
-        const makeAndModel = await certGenSvc.getVehicleMakeAndModel(
-          testResultMock,
-        );
-        expect(makeAndModel.Make).toBe('Isuzu');
-        expect(makeAndModel.Model).toBe('FM');
-      });
-    });
   });
 
   describe('getOdometerHistory function', () => {
